backend/requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.2
aiosignal==1.3.1
amqp==5.2.0
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
autobahn==24.4.2
Automat==22.10.0
bcrypt==4.1.3
billiard==4.2.0
bokeh==3.5.1
Brotli==1.1.0
celery==5.4.0
celery-progress==0.4
certifi==2024.7.4
cffi==1.17.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
contourpy==1.2.1
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.3
cryptography==43.0.0
cssselect2==0.7.0
daphne==4.1.2
defusedxml==0.7.1
diff-match-patch==20230430
distro==1.9.0
Django==5.0.8
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-cors-headers==4.4.0
django-dynamic-model==0.3.0
django-dynamic-models==0.0.5
django-elasticsearch-dsl==8.0
django-environ==0.12.0
django-extensions==3.2.3
django-filter==24.3
django-import-export==4.1.1
django-jazzmin==3.0.0
django-jsonresponse==0.10.0
django-mutant==0.2.1
django-nine==0.2.7
django-picklefield==3.2
django-polymodels==1.8.0
django-redis==5.4.0
django-stubs==5.2.0
django-stubs-ext==5.2.0
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djangorestframework-stubs==3.16.0
dnspython==2.6.1
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.7.1
drf-yasg==1.21.7
elastic-transport==8.13.1
elasticsearch==8.14.0
elasticsearch-dsl==8.14.0
email_validator==2.2.0
et-xmlfile==1.1.0
exceptiongroup==1.2.2
fastapi==0.111.1
fastapi-cli==0.0.5
fonttools==4.56.0
frozenlist==1.4.1
gevent==25.5.1
graphene==3.3
graphene-django==3.2.2
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==3.2.2
gunicorn==22.0.0
h11==0.14.0
hdbcli==2.21.28
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.4
jiter==0.5.0
jsonfield==3.1.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.4.0
loguru==0.7.2
markdown-it-py==3.0.0
MarkupPy==1.14
MarkupSafe==2.1.5
mdurl==0.1.2
msgpack==1.0.8
multidict==6.0.5
numpy==2.0.1
odfpy==1.4.1
openai==1.40.2
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pluggy==1.5.0
promise==2.3
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.4
pydantic-settings==2.3.4
pydantic_core==2.18.4
pydyf==0.11.0
pyfilebrowser==0.1.2
Pygments==2.18.0
pyhdb==0.3.4
PyJWT==2.9.0
pyodbc==5.1.0
pyOpenSSL==24.2.1
pyphen==0.17.2
pyspnego==0.11.1
pytest==8.3.2
pytest-asyncio==0.23.8
pytest-django==4.8.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-multipart==0.0.12
pytz==2024.1
PyYAML==6.0.2
redis==5.0.8
redis-py-cluster==1.3.5
referencing==0.35.1
requests==2.32.3
rich==13.9.1
rpds-py==0.20.0
service-identity==24.1.0
setuptools==72.1.0
shellingham==1.5.4
six==1.16.0
smbprotocol==1.14.0
sniffio==1.3.1
somepackage==1.2.3
SQLAlchemy==2.0.32
sqlalchemy-hana==2.5.0
sqlparse==0.5.1
starlette==0.37.2
tablib==3.5.0
tblib==3.1.0
text-unidecode==1.3
tinycss2==1.4.0
tinyhtml5==2.0.0
tomli==2.0.1
tornado==6.4.1
tqdm==4.66.5
Twisted==24.3.0
txaio==23.1.1
typer==0.12.5
types-PyYAML==6.0.12.20250516
types-requests==2.32.4.20250611
typing_extensions==4.12.2
tzdata==2024.1
ua-parser==0.18.0
uritemplate==4.1.1
urllib3==2.2.2
user-agents==2.2.0
uvicorn==0.30.6
uvloop==0.20.0
vine==5.1.0
watchfiles==0.24.0
wcwidth==0.2.13
weasyprint==64.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==13.1
win32-setctime==1.1.0
xlrd==2.0.1
xlwt==1.3.0
xyzservices==2024.6.0
yarl==1.9.4
zope.event==5.0
zope.interface==7.0.1
zopfli==0.2.3.post1
